[Chore] 코드 수정 #103
